+++ backend/apid/apid.go
@@ -58,6 +58,7 @@ type Option func(*APId) error
 type Config struct {
 	ListenAddress       string
 	RequestLimit        int64
+	WriteTimeout        time.Duration
 	URL                 string
 	Bus                 messaging.MessageBus
 	Store               store.Store
@@ -113,7 +114,7 @@ func New(c Config, opts ...Option) (*APId, error) {
 	a.HTTPServer = &http.Server{
 		Addr:         c.ListenAddress,
 		Handler:      router,
-		WriteTimeout: 15 * time.Second,
+		WriteTimeout: c.WriteTimeout,
 		ReadTimeout:  15 * time.Second,
 		TLSConfig:    tlsServerConfig,
 	}
@@ -231,9 +232,20 @@ func GraphQLSubrouter(router *mux.Router, cfg Config) *mux.Router {
 		middlewares.SimpleLogger{},
 	)
 
+	// The write timeout hangs up the request making it more difficult for
+	// clients to determine what occurred. As such give the service as much time
+	// as possible to produce results.
+	timeout := cfg.WriteTimeout - (50 * time.Millisecond)
+	if timeout < 0 {
+		timeout = 0
+	}
+
 	mountRouters(
 		subrouter,
-		&routers.GraphQLRouter{Service: cfg.GraphQLService},
+		&routers.GraphQLRouter{
+			Service: cfg.GraphQLService,
+			Timeout: timeout,
+		},
 	)
 
 	return subrouter