+++ backend/keepalived/keepalived.go
@@ -725,7 +725,7 @@ func (k *Keepalived) handleUpdate(e *corev2.Event) error {
 				"entity":       entity.Name,
 				"namespace":    entity.Namespace,
 				"subscription": sub,
-				"timeout":      time.Duration(e.Check.Timeout) * time.Second,
+				"timeout":      time.Duration(e.Check.Timeout),
 			})
 			if err := ring.Add(tctx, entity.Name, int64(e.Check.Timeout)); err != nil {
 				lager.WithError(err).Error("error adding entity to ring")