+++ graphql/tracing/prometheus.go
@@ -20,8 +20,9 @@ const (
 var (
 	Collector = prometheus.NewSummaryVec(
 		prometheus.SummaryOpts{
-			Name: "graphql_duration_seconds",
-			Help: "Time spent in GraphQL operations, in seconds",
+			Name:       "graphql_duration_seconds",
+			Help:       "Time spent in GraphQL operations, in seconds",
+			Objectives: map[float64]float64{0.5: 0.05, 0.9: 0.01, 0.99: 0.001},
 		},
 		[]string{"key", "platform_key"},
 	)